Extract welcome controller into controllers/welcome.rs
