+++ slint-ui/src/controllers/mod.rs
@@ -1,0 +1,1 @@
+pub mod welcome;
